Reinserted Meta-CVS code.
When removing the code I forgot to purge. I have done this now.
I didn't know how to re-activate the removed files. They were still in
the MCVS-directory and I could issue mcvs restore but then I had
lots of F-files in lost+found and did not know their original names.
So I just added the files again. If someone knows a better solution to this
problem, please tell me (or is there a feature missing?).