We have a very basic webpage now. It describes the project goal (which was more or less taken out of the scratchpad), provides a link to our development documentation and leaves a note about our repository.
Feel free to add an change whatever you consider necessary.

I have included a Makefile to push up new versions of the pages.
I thought about putting it into some commit-automatism but this does
is not so easily possible because there is no Meta-CVS at berlios.
The Makefile is very basic and will have to be improved in the future.