The new file sbcl-unix.lisp is a replacement for clisp-unix.lisp and cmucl-unix.lisp to use Meta-CVS with SBCL. It is not tested yet. It will not work at the moment because the build-system still has to be modified. But I wanted to have this state checked in anyway.
There also have to be some changes done to dirwalk.lisp because
there are no dir-handles in the Unix glue code anymore.

The new file sbcl-unix.lisp used mostly portable code but some
parts are SBCL-specific. It needs osicat in a patched version which
is in our MCVS (module osicat-patch) and port-path which is a tiny
library based in Peter Seibel's pathnames with smaller functionality
and some extra macros an functions (also in our MCVS, module port-path).
I am sure this will be extended in the next weeks.