Meta-CVS now compiles with SBCL. But the resulting program is not tested yet and very likely to be awfully error prone (and the whole thing took much longer than expected).
There is a basic ASDF definition in ASDF-INSTALL structure
for Meta-CVS and the code can be loaded into SBCL with

(asdf:operate 'asdf:load-op :gennf)

provided ASDF finds gennf.asd in its search paths. The easiest way to
do this is to link it into ~/.sbcl/systems:

$ ln ~/ti/ossi/gennf/src/gennf.asd ~/.sbcl/systems/gennf.asd

The patched version of osicat and port path have to be available as well.

In order to get things working with ASDF all files wer put into package
gennf. Some oddities remain in packages.lisp (see FIXME there).

A build and install script is still missing. This will be the next thing
to do. Up to now the ASDF definition is only suitable for SBCL, for CLISP
or CMUCL (if that works at all) the old build environment has to be
used.