It is now possible to start Meta-CVS as an SBCL image. Currently I work on a usable Makefile.
There were problems with the terminal handling. In the original
Meta-CVS code ctermid was queried and opened as a stream. This
causes trouble with SBCL, probably due to buffering issues.
It is much easier to use *debug-io* as input-/output-stream
for error handling (this way ctermid is not needed and to me
it seems quite reasonable that Common Lisp implementation bind
this variable to something sensible).