Lots of bugfixes concerning the SBCL port. The SBCL build was successful used to do a
$ gennf create

but there remains one stupid problem:

The sb-ext:run-program on Linux with SBCL v0.9.0.18 has trouble
with in- and output for visual programs like vi. On Solaris with
SBCL v0.8.18 it works. I suspect that it is a SBCL bug (as 0.9.0.18
is not the newest version) but this has to be investigated. I think I
try a newer SBCL version but 0.9.6 or something had trouble wih SLIME.