We have now a two-way merge algorithm implemented. Therefore, we have file merging.lisp which performs a two-way merge either on files or sequences. In particular, the following happens: (two-way-merge file1 file2) means that file1 is merged into file2. A diff of the two files is computed using cl-difflib (an implementation of the Gestalt Pattern Matching algorithm by Ratcliff and Obershelp). All chunks of file1 that can be inserted into file2 are inserted. All chunks of file1 that conflict with a chunk in file2 are insrted annotated. All chunks of file2 that do not exist in file1 are inserted with an annotation. The annotations are produced by functions which can be passed to two-way-merge. Furthermore, it is possible to tell two-way-merge, if it shall produce a merge despite conflicts or not. If the corresponding argument is nil, a merge is only produced if no conflicts arise. The return value of two-way-merge is a multiple-value with the first being the merge and the second indicating if conflicts arose or not.
I am not sure, if I broke some parts of the code today, may be it is
inconsistent after this commit.

Next step is probably three-way-merge. This is due to the fact, that on
sandbox update or on commits, there is in fact an ancestor which can
be used during merge. This can in fact be handled by cvs but I think it is
better that all mergeing is either done by cvs or by gennf. And for
branching/merging it cannot be done by cvs, so it should be done by gennf
for update and commit as well.